% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scb.R
\name{get_scb}
\alias{get_scb}
\title{Download SCB data}
\usage{
get_scb(
  file = NA,
  query = NA,
  download_data = TRUE,
  save_data = FALSE,
  url = NA,
  returnList = FALSE
)
}
\arguments{
\item{file}{File name of JSON query}

\item{query}{JSON search query as string.}

\item{download_data}{Boolean indicating whether downloaded or saved data should be used}

\item{save_data}{Boolean indicating whether downloaded data should be saved to a JSON data file}

\item{url}{Url string if not a url field in the JSON query}
}
\value{
A data_frame with time as the first column
}
\description{
\code{get_scb} downloads a JSON query to SCB as a data_frame
}
\details{
For more info, see:

 \url{http://www.github.com/bjornerstedt/scbapi}
}
\examples{
df <- get_scb("gdpreal.json")
df <- get_scb("sick1", colnames = c("time", "sm", "sw"))

}
