% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scb2.R
\name{get_scb2}
\alias{get_scb2}
\title{Download SCB data}
\usage{
get_scb2(fname, download_data = TRUE, save_data = FALSE, url = NA,
  colnames = NULL, returnList = FALSE)
}
\arguments{
\item{fname}{File name of JSON query}

\item{download_data}{Boolean indicating whether downloaded or saved data should be used}

\item{save_data}{Boolean indicating whether downloaded data should be saved to a JSON data file}

\item{url}{Url string if not a url field in the JSON query}

\item{colnames}{String array of column names.}
}
\value{
A data_frame with time as the first column
}
\description{
\code{get_scb} downloads a JSON query to SCB as a data_frame
}
\details{
For more info, see:

\url{http://www.github.com/bjornerstedt/scbapi}
}
\examples{
df <- get_scb("gdpreal.json")
df <- get_scb("sick1", colnames = c("time", "sm", "sw"))

}

