% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scb_api.R
\name{scb_api}
\alias{scb_api}
\title{Query SCB API}
\usage{
scb_api(id = "ssd")
}
\arguments{
\item{id}{= url string}
}
\description{
Build query for scb api
}
\examples{
a <- scb_api("ssd/BE/BE0001/BE0001T100")
res <- a$get()
}

