% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scb.R
\name{scb_data}
\alias{scb_data}
\title{\code{scb_data} Converts a JSON dataset from SCB to a data_frame}
\usage{
scb_data(file)
}
\arguments{
\item{file}{File name of JSON query}
}
\description{
For more info, see:
}
\details{
\url{http://www.github.com/bjornerstedt/scbapi}
}
\examples{
df <- scb_data("gdpreal.json")

}
